requirements.txt
langchain
pandas
openai
python-dotenv
streamlit
pymongo